# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Library",
        "description": "Library NodeJS Application API",
        "licence": "ISC"
    },
    "basePath": "/api/v1",
    "tags": [
        {
            "name": "Book",
            "description": "API for book in the system"
        }
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/book/create": {
            "post":{
                "tags": [
                    "Book"
                ],
                "description": "Create new book in system",
                "parameters": [
                    {
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "description": "Book that we want to create",
                        "example":{
                            "title": "Cien años de soledad",
                            "pages": 402,
                            "year": 1964,
                            "isbn": "123-456-7890-A1",
                            "author": "Gabriel García Márquez"
                        },
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }

                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "New Book is created",
                        "examples": {
                            "application/json": {
                                "status": true,
                                "message": "Book saved!"
                            }
                        }

                    }
                }
            }
        },
        "/book/list":{
            "get": {
                "tags": [
                    "Book"
                ],
                "responses": {
                    "200": {
                        "description": "All books are returned",
                        "examples": {
                            "application/json":{
                                "status": true,
                                "listbooks": "array"
                            }
                        }
                    }
                }
            }
        },

        "/book/{isbn}":{
            "get":{
                "tags": [
                    "Book"
                ],
                "responses":{
                    "200": {
                        "description": "A book returned",
                        "examples":{
                            "application/json": {
                                "status": true,
                                "bookByISBN": "object"
                            }
                        }
                    }
                }
            },
            "parameters":[
                {
                    "name": "isbn",
                    "in": "path",
                    "required": false,
                    "description": "Book isbn to find",
                    "type": "string"
                }
            ]
        },
        "/book/delete/{isbn}":{
            "delete":{
                "tags": [
                    "Book"
                ],
                "responses":{
                    "200": {
                        "description": "A book deleted",
                        "examples":{
                            "application/json": {
                                "status": true,
                                "message": "Book deleted"
                            }
                        }
                    }
                }
            },
            "parameters":[
                {
                    "name": "isbn",
                    "in": "path",
                    "required": false,
                    "description": "Book isbn to delete",
                    "type": "string"
                }
            ]
            
        },
        "/book/update/{isbn}":{
            "put":{
                "tags": [
                    "Book"
                ],
                "responses":{
                    "200": {
                        "description": "A book updated",
                        "examples":{
                            "application/json": {
                                "status": true,
                                "message": "Book updated!"
                            }
                        }
                    }
                },
                "parameters":[
                    {
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "description": "Book to update",
                        "schema": {
                            "$ref": "#/definitions/Book"
                        }
                    }
                ]
            },
            "parameters":[
                {
                    "name": "isbn",
                    "in": "path",
                    "required": false,
                    "description": "Book isbn to update",
                    "type": "string"
                }
            ]
            
            
        }


    },
    "definitions":{
        "Book": {
            "required":[
                "title",
                "pages",
                "year",
                "isbn",
                "author"
            ],
            "properties":{
                "title": {
                    "type": "string",
                    "uniqueItems": false,
                    "example": "Cien años de Soledad"
                },
                "pages": {
                    "type": "integer",
                    "example": 402
                },
                "year": {
                    "type": "integer",
                    "example": 1962
                },
                "isbn": {
                    "type": "string",
                    "uniqueItems": false,
                    "example": "123-456-7890-A1"
                },
                "author": {
                    "type": "string",
                    "uniqueItems": false,
                    "example": "Gabriel García Márquez"
                }
            }
        }
    }
}